# api/docs/swagger.yaml
definitions:
  common.PageResponse:
    properties:
      next_key:
        items:
          type: integer
        type: array
        x-order:0: true
      total:
        type: integer
        x-order:1: true
    type: object
  nft.Collection:
    properties:
      address:
        type: string
        x-order:1: true
      creator:
        type: string
        x-order:0: true
      name:
        type: string
        x-order:2: true
      nft_count:
        type: integer
        x-order:4: true
      origin_name:
        type: string
        x-order:3: true
    type: object
  nft.CollectionsResponse:
    properties:
      collections:
        items:
          $ref: '#/definitions/nft.Collection'
        type: array
        x-order:0: true
      pagination:
        allOf:
        - $ref: '#/definitions/common.PageResponse'
        x-order:1: true
    type: object
  nft.Nft:
    properties:
      collection_address:
        type: string
        x-order:0: true
      nft_token_id:
        type: string
        x-order:2: true
      object_addr:
        description: only used in Move
        type: string
        x-order:1: true
      owner:
        type: string
        x-order:3: true
      uri:
        type: string
        x-order:4: true
    type: object
  nft.NftsResponse:
    properties:
      pagination:
        allOf:
        - $ref: '#/definitions/common.PageResponse'
        x-order:1: true
      tokens:
        items:
          $ref: '#/definitions/nft.Nft'
        type: array
        x-order:0: true
    type: object
...
info:
  contact: {}
paths:
...
  /indexer/block/v1/blocks:
    get:
      consumes:
      - application/json
      description: Get a list of blocks with pagination
      parameters:
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses: {}
      summary: Get blocks
      tags:
      - Blocks
  /indexer/block/v1/blocks/{height}:
    get:
      consumes:
      - application/json
      description: Get a specific block by its height
      parameters:
      - description: Block height
        in: path
        name: height
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      summary: Get block by height
      tags:
      - Blocks
  /indexer/nft/v1/collections:
    get:
      consumes:
      - application/json
      description: Get NFT collections for a specific collection and token ID
      parameters:
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/nft.CollectionsResponse'
      summary: Get NFT collections by collection and token ID
      tags:
      - NFT
  /indexer/nft/v1/collections/{address}:
    get:
      consumes:
      - application/json
      description: Get NFT collections for a specific address
      parameters:
      - description: Collection address
        in: path
        name: address
        required: true
        type: string
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/nft.CollectionsResponse'
      summary: Get NFT Collections By Collection Address
      tags:
      - NFT
  /indexer/nft/v1/collections/by_account/{account}:
    get:
      consumes:
      - application/json
      description: Get NFT collections owned by a specific account
      parameters:
      - description: Account address
        in: path
        name: account
        required: true
        type: string
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/nft.CollectionsResponse'
      summary: Get NFT collections by account
      tags:
      - NFT
  /indexer/nft/v1/collections/by_name/{name}:
    get:
      consumes:
      - application/json
      description: Get NFT collections for a specific name
      parameters:
      - description: Collection name
        in: path
        name: name
        required: true
        type: string
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/nft.CollectionsResponse'
      summary: Get NFT collections by name
      tags:
      - NFT
  /indexer/nft/v1/tokens/by_account/{account}:
    get:
      consumes:
      - application/json
      description: Get NFT tokens owned by a specific account
      parameters:
      - description: Account address
        in: path
        name: account
        required: true
        type: string
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/nft.NftsResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get NFT tokens by account
      tags:
      - NFT
  /indexer/nft/v1/tokens/by_collection/{collection_addr}:
    get:
      consumes:
      - application/json
      description: Get NFT tokens from a specific collection
      parameters:
      - description: Collection address
        in: path
        name: collection_addr
        required: true
        type: string
      - description: Pagination key
        in: query
        name: pagination.key
        type: string
      - description: Pagination offset
        in: query
        name: pagination.offset
        type: integer
      - default: 100
        description: Pagination limit
        in: query
        name: pagination.limit
        type: integer
      - description: Count total
        in: query
        name: pagination.count_total
        type: boolean
      - description: Reverse order
        in: query
        name: pagination.reverse
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/nft.NftsResponse'
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get NFT tokens by collection
      tags:
      - NFT
...
swagger: "2.0"